Attendance_record insertion

+++ src/main/resources/db/migration/V6__Create_Table_Attendance_Record.sql
@@ -2,7 +2,7 @@ create table IF NOT EXISTS attendance_record(
     date date not null,
     subject_id varchar(10),
     student_id varchar(10),
-    presence int,
+    presence tinyint(1),
     FOREIGN KEY (subject_id) REFERENCES subject(id),
     FOREIGN KEY (student_id) REFERENCES student(id)
 );
